requirements.txt
rumps>=0.4.0
mss>=9.0.1
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0
pyautogui>=0.9.54
python-dateutil>=2.8.0
# pyobjc-framework-AppKit - install separately if needed: pip install pyobjc-framework-AppKit
# pyobjc-framework-EventKit - install separately if needed: pip install pyobjc-framework-EventKit
# pyobjc-framework-UserNotifications - install separately if needed: pip install pyobjc-framework-UserNotifications
# Note: UserNotifications framework is used for notification permissions

